DPC-80: Refactored Hibernate configuration/injection into its own module.
We can now inject the SessionFactory and Sessions using a standalone module.
This module is now a part of dpc-common, so it should be easy to inject.
