+++ dpc-attribution/pom.xml
@@ -66,11 +66,6 @@
             <artifactId>dpc-common</artifactId>
             <version>0.2.0-SNAPSHOT</version>
         </dependency>
-        <dependency>
-            <groupId>org.reflections</groupId>
-            <artifactId>reflections</artifactId>
-            <version>0.9.11</version>
-        </dependency>
         <!--Testing dependencies-->
         <dependency>
             <groupId>io.dropwizard</groupId>
