+++ dpc-attribution/src/main/resources/application.conf
@@ -1,7 +1,7 @@
 dpc.attribution {
   database {
     driverClass = org.postgresql.Driver
-    url = "jdbc:postgresql://localhost:5432/dpc_attribution"
+    url = "jdbc:postgresql://localhost:5432/dpc-dev"
     user = postgres
   }
