+++ dpc-common/src/main/java/gov/cms/dpc/common/hibernate/IDPCDatabase.java
@@ -1,0 +1,7 @@
+package gov.cms.dpc.common.hibernate;
+
+import io.dropwizard.db.DataSourceFactory;
+
+public interface IDPCDatabase {
+    DataSourceFactory getDatabase();
+}
